third_party/blink/web_tests/html5lib/generated/run-webkit01-data-expected.txt

CONSOLE MESSAGE: PASS
CONSOLE MESSAGE: FOO<span>BAR</span>BAZ
../resources/webkit01.dat: PASS
